+++ core/src/api/index.ts
@@ -1,5 +1,3 @@
-export * from './auth'
-export * from './course'
 export * from './notification'
 export * from './score'
 export * from './song'
