--- functions/getSongs/index.test.ts
@@ -1,17 +1,0 @@
-import { testSongData } from '@ddradar/core/__tests__/data'
-import { describe, expect, test } from 'vitest'
-
-import getSongs from '.'
-
-describe('GET /api/v1/songs', () => {
-  test(`returns "200 OK" with JSON`, async () => {
-    // Arrange
-    const songs = [{ ...testSongData }]
-
-    // Act
-    const result = await getSongs(null, null, songs)
-
-    // Assert
-    expect(result).toStrictEqual({ status: 200, body: songs })
-  })
-})
