+++ web/server/auth.ts
@@ -1,10 +1,35 @@
-import type { Api, Database } from '@ddradar/core'
+import type { Database } from '@ddradar/core'
 import { fetchLoginUser } from '@ddradar/db'
 import type { CompatibilityEvent } from 'h3'
 
+type Role = 'anonymous' | 'authenticated' | 'administrator'
+
+/**
+ * User information provided by Azure
+ * @see https://docs.microsoft.com/azure/static-web-apps/user-information
+ */
+export interface ClientPrincipal  {
+  /** The name of the identity provider. */
+  identityProvider: 'github' | 'twitter'
+  /** An Azure Static Web Apps-specific unique identifier for the user. */
+  userId: string
+  /** User Name (GitHub/Twitter) */
+  userDetails: string
+  /**
+   * An array of the user's assigned roles.
+   * @see https://docs.microsoft.com/azure/static-web-apps/authentication-authorization#roles
+   */
+  userRoles: ReadonlyArray<Role>
+}
+
+/**
+ * Get {@link ClientPrincipal} from Request header.
+ * @param event HTTP Event
+ * @description https://docs.microsoft.com/azure/static-web-apps/user-information?tabs=javascript#api-functions
+ */
 export function useClientPrincipal(
   event: Pick<CompatibilityEvent, 'req'>
-): Api.ClientPrincipal | null {
+): ClientPrincipal | null {
   const header = event.req.headers['x-ms-client-principal']
   if (typeof header !== 'string') return null
 
@@ -18,7 +43,7 @@ export function useClientPrincipal(
 }
 
 export async function getLoginUserInfo(
-  clientPrincipal: Pick<Api.ClientPrincipal, 'userId'> | null
+  clientPrincipal: Pick<ClientPrincipal, 'userId'> | null
 ): Promise<Database.UserSchema | null> {
   if (!clientPrincipal) return null
   return await fetchLoginUser(clientPrincipal.userId)
